employee assign tour, check error

+++ project/employee_delete.php
@@ -5,8 +5,31 @@ session_start();
 $user_id = isset($_GET['user_id']) ? $_GET['user_id'] :  die('ERROR: Record ID not found.');
 include 'config/database.php';
 
+// user's ID stored in the session. 
 try {
+    $check_user = "SELECT employee.user_id, employee.username
+    FROM employee
+    INNER JOIN `LEAVE` ON employee.user_id = `LEAVE`.user_id
+    WHERE employee.user_id = 104;
+    ";
+    $stmt = $con->prepare($check_user);
+    $stmt->bindParam(":user_id", $user_id);
+    $stmt->execute();
+    $row = $stmt->fetch(PDO::FETCH_ASSOC);
+    $count = $stmt->rowCount();
 
+    $flag =0;
+    if ($count > 0) {
+        header('Location: employee_read.php?action=faildelete');
+        $flag =1;
+    }
+
+    if ($_SESSION['user_id'] == $user_id) {
+        header('Location: employee_read.php?action=faildeleteown');
+        $flag =1;
+    }
+
+    if ($flag == 0) {
         // delete query
         $query = "DELETE FROM employee WHERE user_id = ?";
         $stmt = $con->prepare($query);
@@ -21,9 +44,10 @@ try {
             }
             header('Location: employee_read.php?action=deleted');
         } else {
-            die('Unable to delete record.');
+            header('Location: employee_read.php?action=faildelete');
         }
     }
+}
 
 // show error
 catch (PDOException $exception) {
